Allgemeinen HPController hinzugefügt
