+++ Platformer_001/Assets/Scripts/HPController.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 833aeb026f2d4e041aa8bfd7fa284160
-timeCreated: 1493804739
+guid: 2145912209ceed54a836f55a6431ad43
+timeCreated: 1493804540
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
