+++ Platformer_001/Assets/Scripts/SpielerHP.cs
@@ -1,68 +1,20 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.Events;
-using UnityEngine.EventSystems;
+﻿using UnityEngine;
 
-public class SpielerHP : MonoBehaviour {
-	[Header("Werte")]
-	[Range(0, 100)]
-	public int hp = 100;
-	public int maxHP = 100;
-
-	[Header("Events")]
-    public EventTrigger.TriggerEvent schadenEvent;
-    public EventTrigger.TriggerEvent todEvent;
-
-    public HUDHPAnzeige hpAnzeige;
-
-    void Start() {
-        if (hpAnzeige != null) hpAnzeige.HPSetzen(hp);
-    }
-
-	private void OnSterben()
-	{
-		todEvent.Invoke(null);
+public class SpielerHP : HPController {
+	public override void OnSterben() {
+		base.OnSterben();
 
 		var leben = gameObject.GetComponent<LebenController> ();
-		if (leben != null) 
-		{
-			leben.Sterben ();
+		if (leben != null) {
+			leben.Sterben();
 		}
 
-		var levelGen = FindObjectOfType<LevelGenerator> ();
+		var levelGen = FindObjectOfType<LevelGenerator>();
 		if (levelGen != null) {
-			levelGen.NeuGenerieren ();
+			levelGen.NeuGenerieren();
 			Vector3 newPos = new Vector3 (levelGen.spawnPoint.x, levelGen.spawnPoint.y, 0.0f);
 			transform.position = newPos;
 			hp = 100;
 		}
-
-		var endlosGen = FindObjectOfType<Endlosmodus> ();
-		if (endlosGen != null) {
-			// TODO: Spawnpunkt des Endlosmodus finden??
-		}
-	}
-
-	public void SchadenZufuegen(int schaden) {
-        hp = Math.Max(hp - schaden, 0);
-
-        if (hp <= 0) {
-			OnSterben ();
-        } else {
-            schadenEvent.Invoke(null);
-        }
-
-        if (hpAnzeige != null) hpAnzeige.HPSetzen(hp);
-    }
-
-    public void Heilen(int hp) {                        
-        this.hp = hp + Math.Min(maxHP, hp);
-    }
-
-	public void Umbringen()
-	{
-		SchadenZufuegen (hp);
 	}
 }
